% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecb_fx.R
\name{get_ecb_fx}
\alias{get_ecb_fx}
\title{Get Exchange Rate vs EUR from ECB}
\usage{
get_ecb_fx(
  cur,
  start_date = "1999-01-04",
  end_date = Sys.Date(),
  freq = "D",
  type = "A"
)
}
\arguments{
\item{cur}{Currency - three letter standard}

\item{start_date}{First day of time series}

\item{end_date}{Last day of time series}

\item{freq}{Upper case, single letter abbreviation of annual, daily, half-yearly, monthly or quarterly}

\item{type}{'A' for average or 'E' for end-of-period}
}
\value{
A tibble with Datum as date index and y.xxx.EUR as values for currency XXX with frequency y
}
\description{
This function implements the API to the \href{https://sdw.ecb.europa.eu/home.do}{ECB Statistical Data Warehouse} to
fetch reference exchange rates vs the Euro.
}
\details{
The type "end-of-period" does not exist for
frequencies (e. g. half-annually for USD). If this is the case, use "A" for average.
The "Datum" column is only of type "date" for daily data, else it is "character".
}
\examples{
get_ecb_fx("CHF")
}
