% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecb_fx.R
\name{get_ecb_fx}
\alias{get_ecb_fx}
\title{Get Exchange Rate vs EUR from ECB}
\usage{
get_ecb_fx(
  cur,
  start_date = "1999-01-04",
  end_date = Sys.Date(),
  freq = "D",
  type = "A"
)
}
\arguments{
\item{cur}{Currency - three letter standard}

\item{start_date}{First day of time series}

\item{end_date}{Last day of time series}

\item{freq}{Single letter abbreviation of annual, daily, half-yearly, monthly or quarterly}

\item{type}{'A' for average or 'E' for end-of-period}
}
\value{
A tibble with Datum as date index and y.xxx.EUR for currency XXX with frequency y
}
\description{
Get Exchange Rate vs EUR from ECB
}
\examples{
get_ecb_fx("CHF")
}
